Raise BPM to 1000, add adjustable time signature, rework cadence controls
BPM now goes to 1000 (was 300 in the UI, 999 in the clock). "Next every"
extends down to a quarter note via a new Cadence type (absolute note
values vs. bars-of-signature), and time signature is now adjustable
(arbitrary numerator, denominator in 1/2/4/8/16) while BPM stays
quarter-note based, Link-compatible. Persists in .viproj with fallback
to the legacy quantum/phrase_len/loop_len fields for old projects.

Replaces the next-every/loop-every bracket-button rows with a new
scroll-with-detents phosphor widget, also used for the time-sig control.
Went through a few interaction bugs found via live testing: cadence
comparison by resolved ticks instead of enum variant (fixes a control
showing no selection), and the widget ended up supporting left-click
(+1), right-click (-1), and click-drag scrubbing (plus wheel), since a
click-only affordance didn't match how the rest of this UI responds.

+++ src/sequencer.rs
@@ -244,8 +244,8 @@ impl Sequencer {
     /// Change the global default dwell (the "next every" phrase length). Any
     /// armed cue is disarmed (its fire beat was computed against the old length);
     /// it re-arms at the new window's arm point.
-    pub fn set_phrase_len(&mut self, beats: u32) -> Vec<SequencerEvent> {
-        self.default_dwell = beats.max(1) as f64;
+    pub fn set_phrase_len(&mut self, beats: f64) -> Vec<SequencerEvent> {
+        self.default_dwell = beats.max(1.0);
         if let SeqState::PlayingArmed { cue, started, .. } = self.state {
             self.state = SeqState::Playing { cue, started };
             return vec![SequencerEvent::DisarmDecoder];
